+++ common/src/main/java/net/hibiscus/naturespirit/blocks/GrowingBranchingTrunkBlock.java
@@ -2,7 +2,7 @@ package net.hibiscus.naturespirit.blocks;
 
 
 import net.hibiscus.naturespirit.registration.NSTags;
-import net.minecraft.advancements.CriteriaTriggers;
+import net.minecraft.advancements.triggers.CriteriaTriggers;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;