+++ common/src/main/java/net/hibiscus/naturespirit/registration/NSRegistryHelper.java
@@ -8,6 +8,7 @@ import net.hibiscus.naturespirit.registration.compat.NSArtsAndCraftsCompat;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.EntityTypes;
 import net.minecraft.world.item.DoubleHighBlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.BlockGetter;
@@ -55,7 +56,7 @@ public class NSRegistryHelper {
     }
 
     public static Boolean ocelotOrParrot(BlockState state, BlockGetter world, BlockPos pos, EntityType<?> type) {
-        return type == EntityType.OCELOT || type == EntityType.PARROT;
+        return type == EntityTypes.OCELOT || type == EntityTypes.PARROT;
     }
 
     public static <T extends Block> NSBlockHolder<T> registerBlockWithoutItem(String name, Function<BlockBehaviour.Properties, ? extends T> block, Supplier<BlockBehaviour.Properties> properties) {