+++ common/src/main/java/net/hibiscus/naturespirit/world/carver/ReplaceableRavineCarver.java
@@ -105,7 +105,7 @@ public class ReplaceableRavineCarver extends WorldCarver<ReplaceableRavineCarver
       } else {
         chunk.setBlockState(pos, blockState2);
         if (aquiferSampler.shouldScheduleFluidUpdate() && !blockState2.getFluidState().isEmpty()) {
-          chunk.markPosForPostprocessing(pos);
+          chunk.markPosForPostProcessing(pos);
         }
 
         if (replacedGrassy.isTrue()) {
@@ -114,7 +114,7 @@ public class ReplaceableRavineCarver extends WorldCarver<ReplaceableRavineCarver
             context.topMaterial(posToBiome, chunk, tmp, !blockState2.getFluidState().isEmpty()).ifPresent((state) -> {
               chunk.setBlockState(tmp, state);
               if (!state.getFluidState().isEmpty()) {
-                chunk.markPosForPostprocessing(tmp);
+                chunk.markPosForPostProcessing(tmp);
               }
 
             });